requirements.txt
sphinx
livereload
myst-parser
sphinx-autodoc-typehints
mongoengine
python-dotenv
dnspython
pymongo